move public key
