+++ cmd/mailchain/internal/rest/handlers/publickey.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package publickey
+package handlers
 
 import (
 	"encoding/json"
